fix: test case: member top role info
